Changning %used from int to float

+++ collectors/0/dfstat.py
@@ -111,7 +111,7 @@ def main():
         continue
 
       used = r.f_blocks - r.f_bfree
-      percent_used = 100 if r.f_blocks == 0 else int(used / float(r.f_blocks) * 100)
+      percent_used = 100 if r.f_blocks == 0 else used / float(r.f_blocks) * 100
       print("df.bytes.total %d %s mount=%s fstype=%s"
             % (ts, r.f_frsize * r.f_blocks, fs_file, fs_vfstype))
       print("df.bytes.used %d %s mount=%s fstype=%s"
@@ -122,7 +122,7 @@ def main():
             % (ts, r.f_frsize * r.f_bfree, fs_file, fs_vfstype))
 
       used = r.f_files - r.f_ffree
-      percent_used = 100 if r.f_files == 0 else int(used / float(r.f_files) * 100)
+      percent_used = 100 if r.f_files == 0 else used / float(r.f_files) * 100
       print("df.inodes.total %d %s mount=%s fstype=%s"
             % (ts, r.f_files, fs_file, fs_vfstype))
       print("df.inodes.used %d %s mount=%s fstype=%s"
